+++ src/db/model_oblique_strategy.rs
@@ -1,21 +1,17 @@
-use serde::Deserialize;
 use sqlx::{Sqlite, SqlitePool, Transaction};
 
 use crate::app_error::AppError;
 
-#[derive(sqlx::FromRow, Debug, Clone, Deserialize, PartialEq, Eq, PartialOrd, Ord)]
-pub struct ModelObliqueStratergy {
-    pub stratergy: String,
-}
+pub struct ModelObliqueStrategy;
 
-impl ModelObliqueStratergy {
+impl ModelObliqueStrategy {
     async fn insert(
         transaction: &mut Transaction<'_, Sqlite>,
-        stratergy: &str,
+        trategy: &str,
     ) -> Result<(), AppError> {
-        let sql = "INSERT INTO oblique_strategies (stratergy) VALUES($1) ON CONFLICT(stratergy) DO NOTHING";
+        let sql = "INSERT INTO oblique_strategies (strategy) VALUES($1) ON CONFLICT(strategy) DO NOTHING";
         sqlx::query(sql)
-            .bind(stratergy)
+            .bind(trategy)
             .execute(&mut **transaction)
             .await?;
         Ok(())
@@ -24,8 +20,8 @@ impl ModelObliqueStratergy {
     pub async fn seed_stratergies(sqlite: &SqlitePool) -> Result<(), AppError> {
         let all_stratergies = include_str!("./oblique.txt");
         let mut transaction = sqlite.begin().await?;
-        for stratergy in all_stratergies.lines() {
-            Self::insert(&mut transaction, stratergy).await?;
+        for trategy in all_stratergies.lines() {
+            Self::insert(&mut transaction, trategy).await?;
         }
         transaction.commit().await?;
 
@@ -46,7 +42,7 @@ mod tests {
     use super::*;
 
     #[tokio::test]
-    async fn model_oblique_stratergy_seed() {
+    async fn model_oblique_trategy_seed() {
         let uuid = Uuid::new_v4();
 
         let app_envs = gen_app_envs(uuid);
@@ -55,12 +51,12 @@ mod tests {
         let db = get_db(&app_envs).await.unwrap();
         create_tables(&db).await;
 
-        let result = ModelObliqueStratergy::seed_stratergies(&db).await;
+        let result = ModelObliqueStrategy::seed_stratergies(&db).await;
 
         assert!(result.is_ok());
 
         // Make sure conflicts are ignored, in that they don't cause errors
-        let result = ModelObliqueStratergy::seed_stratergies(&db).await;
+        let result = ModelObliqueStrategy::seed_stratergies(&db).await;
 
         assert!(result.is_ok());
         test_cleanup(uuid, Some(db)).await;